mcp-servers/tala-memory-graph/requirements.txt
mcp>=1.0.0
fastmcp>=0.1.0
pydantic>=2.0
PyYAML
psycopg2-binary>=2.9
